Fix incorrect parsing of auth_results
The DMARC spec allows for zero or more DKIM results and one or more SPF
results.  Prior to this commit, we only allowed for a single result for
each type.  Update the dmarc.AuthResults struct to use a slice for each
type.

For reference, see the definition of AuthResultType in Appendix C of RFC
7489.

This change in the data model required updates to the Go templates.  The
two HTML templates are almost exactly the same, so pull the common body
out into a separate const variable to ease maintenance.

Fixes #45